backend/requirements.txt
# FastAPI 및 서버
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Yahoo Finance API
yahooquery>=2.3.1
yfinance>=0.2.28

# 데이터 처리
pandas>=2.0.0

# AI 및 뉴스
google-genai>=0.2.0

# 유틸리티
python-dotenv>=1.0.0
requests>=2.31.0

# 스케줄러
APScheduler>=3.10.4

# 이미지 처리
Pillow>=10.0.0

# Pydantic (FastAPI에 포함되지만 명시)
pydantic>=2.0.0
